move location import to location package
